add operator= to Task2

+++ Task2/Task2.h
@@ -5,12 +5,12 @@
 
 // Инициализация массива integer
 template<typename T>
-void initMatrix(Table<T>& matrix, size_t raws, size_t columns)
+void initMatrix(Table<T>& matrix, int raws, int columns)
 {
-	size_t count{};
-	for (size_t raw{}; raw < raws; ++raw)
+	int count{};
+	for (int raw{}; raw < raws; ++raw)
 	{
-		for (size_t column{}; column < columns; ++column)
+		for (int column{}; column < columns; ++column)
 		{
 			matrix[raw][column] = static_cast<T>(count++);
 		}
@@ -19,11 +19,11 @@ void initMatrix(Table<T>& matrix, size_t raws, size_t columns)
 
 // Вывод массива в консоль
 template<typename T>
-void printMatrix(Table<T>& matrix, size_t raws, size_t columns)
+void printMatrix(Table<T>& matrix, int raws, int columns)
 {
-	for (size_t raw{}; raw < raws; ++raw)
+	for (int raw{}; raw < raws; ++raw)
 	{
-		for (size_t column{}; column < columns; ++column)
+		for (int column{}; column < columns; ++column)
 		{
 			std::cout << (matrix[raw][column]) << ' ';
 		}
